+++ src/deion/ui/views/Account.tsx
@@ -339,6 +339,7 @@ const Account = ({
 			</>
 		);
 	}
+	console.log("goldSuccess", goldSuccess);
 
 	return (
 		<>