+++ src/deion/worker/views/account.ts
@@ -28,7 +28,7 @@ const updateAccount = async (
 		return {
 			email: partialTopMenu.email,
 			goldMessage: inputs.goldMessage,
-			goldSuccess: !!inputs.goldSuccess,
+			goldSuccess: inputs.goldSuccess,
 			goldUntilDateString,
 			loggedIn,
 			showGoldActive,