+++ src/deion/worker/views/roster.ts
@@ -81,6 +81,7 @@ const updateRoster = async (
 			"watch",
 			"untradable",
 			"hof",
+			"latestTransaction",
 		]; // tid and draft are used for checking if a player can be released without paying his salary
 
 		const ratings = ["ovr", "pot", "dovr", "dpot", "skills", "pos"];